[fix] delete 예외처리 수정
관련 test 코드 수정

+++ src/main/java/shop/bookbom/shop/domain/wish/service/impl/WishServiceImpl.java
@@ -66,8 +66,13 @@ public class WishServiceImpl implements WishService {
     @Transactional
     public void deleteWish(List<WishAddDeleteRequest> items, Long userId) {
         items.forEach(item -> {
-                    Wish wish = wishRepository.findByBookIdAndMemberId(item.getBookId(), userId);
-                    wishRepository.delete(wish);
+                    boolean isExistWish = wishRepository.existsByBookIdAndMemberId(item.getBookId(), userId);
+                    if(!isExistWish){
+                        throw new WishNotFoundException();
+                    } else{
+                        Wish wish = wishRepository.findByBookIdAndMemberId(item.getBookId(), userId);
+                        wishRepository.delete(wish);
+                    }
                 }
         );
     }
